src/main/java/com/codewars/uniquenumber/UniqueFinder.java
package com.codewars.uniquenumber;

import java.util.Arrays;
import java.util.Map;

import static java.util.function.Function.identity;
import static java.util.stream.Collectors.counting;
import static java.util.stream.Collectors.groupingBy;

public class UniqueFinder {
    public double findUnique(double[] numbers) {
        return Arrays.stream(numbers)
                .boxed()
                .collect(groupingBy(identity(), counting()))
                .entrySet().stream()
                .filter(entry -> entry.getValue() == 1)
                .map(Map.Entry::getKey)
                .findFirst()
                .orElse(0.0);
    }
}